pushing current exps trails setups

+++ dotless_arabic/experiments/translation/src/utils.py
@@ -7,7 +7,7 @@ import torch
 
 from tqdm.auto import tqdm
 
-from torchmetrics.text.bleu import BLEUScore
+from torchmetrics.text import BLEUScore, SacreBLEUScore
 
 from pytorch_lightning import Trainer
 from pytorch_lightning.callbacks import (
@@ -42,6 +42,7 @@ def train_translator(
     callbacks=None,
     one_run=False,
     text_type="dotted",
+    validate_and_fit=True,
     use_rich_progressbar=True,
     max_epochs=constants.MAX_EPOCHS,
 ):
@@ -49,7 +50,8 @@ def train_translator(
     checkpoints_path = Path(f"NMT/{text_type}/{tokenizer_class.__name__}/checkpoints")
     if callbacks is None:
         callbacks = []
-    shutil.rmtree(checkpoints_path, ignore_errors=True)
+    if validate_and_fit:
+        shutil.rmtree(checkpoints_path, ignore_errors=True)
     checkpoint_callback = ModelCheckpoint(
         mode="min",
         save_top_k=1,
@@ -93,15 +95,16 @@ def train_translator(
         # log_every_n_steps=max(len(train_dataloader) // 25, 1),
         log_every_n_steps=max(len(train_dataloader) // 25, 1),
     )
-    trainer.validate(
-        model=translator,
-        dataloaders=val_dataloader,
-    )
-    trainer.fit(
-        translator,
-        train_dataloader,
-        val_dataloader,
-    )
+    if validate_and_fit:
+        trainer.validate(
+            model=translator,
+            dataloaders=val_dataloader,
+        )
+        trainer.fit(
+            translator,
+            train_dataloader,
+            val_dataloader,
+        )
     return trainer
 
 
@@ -150,7 +153,8 @@ def get_source_tokenizer(
     tokenizer_class=WordTokenizer,
 ):
     if tokenizer_class == SentencePieceTokenizer:
-        tokenizer = tokenizer_class(vocab_size=10**4)
+        # tokenizer = tokenizer_class(vocab_size=10**4)
+        tokenizer = tokenizer_class(vocab_size=5_000)
     else:
         tokenizer = tokenizer_class(
             vocab_size=10**10
@@ -176,7 +180,8 @@ def get_target_tokenizer(
 ):
     if tokenizer_class == SentencePieceTokenizer:
         tokenizer = tokenizer_class(
-            vocab_size=10**4,
+            # vocab_size=10**4,
+            vocab_size=5_000,
             special_tokens=["<bos>", "<eos>"],
         )
     else:
@@ -186,16 +191,16 @@ def get_target_tokenizer(
         )
     tokenizer.train(text="\n".join(train_dataset[target_language_code]))
     # delete 1-freq words if tokenizer class is not Sentencepiece
-    if tokenizer_class != SentencePieceTokenizer:
-        vocab = {
-            vocab: freq
-            for vocab, freq in tokenizer.vocab.items()
-            if freq > 1 or freq < 0
-        }
-        tokenizer.vocab = vocab
-        tokenizer.vocab_size = len(vocab)
-        tokenizer.vocab_size
-        tokenizer.vocab_size
+    # if tokenizer_class != SentencePieceTokenizer:
+    #     vocab = {
+    #         vocab: freq
+    #         for vocab, freq in tokenizer.vocab.items()
+    #         if freq > 1 or freq < 0
+    #     }
+    #     tokenizer.vocab = vocab
+    #     tokenizer.vocab_size = len(vocab)
+    #     tokenizer.vocab_size
+    #     tokenizer.vocab_size
     return tokenizer
 
 
@@ -209,9 +214,11 @@ def get_blue_score(
     blue_n_gram=4,
     use_tqdm=True,
 ):
+    source_sentences = source_sentences[:30]
+    target_sentences = target_sentences[:30]
     if use_tqdm:
-        target_sentences = tqdm(target_sentences)
         source_sentences = tqdm(source_sentences)
+        target_sentences = tqdm(target_sentences)
     targets = [[f"<bos> {sentence} <eos>"] for sentence in target_sentences]
     predictions = [
         model.translate(
@@ -222,8 +229,23 @@ def get_blue_score(
         )
         for sentence in source_sentences
     ]
+    predictions = list(map(lambda text: text.replace("▁ ", ""), predictions))
+    targets = list(
+        map(
+            lambda texts: [text.replace("▁ ", "") for text in texts],
+            targets,
+        )
+    )
+    for s, p, t in zip(source_sentences, predictions, targets):
+        print(f"Source: {s}")
+        print(f"Prediction: {p}")
+        print(f"Target: {t}")
     blue_calculator = BLEUScore(n_gram=blue_n_gram)
     blue_score = blue_calculator(predictions, targets)
+    print("blue score", blue_score)
+    sacre_bleu_calculator = SacreBLEUScore(ngram=blue_n_gram)
+    sacre_blue_score = sacre_bleu_calculator(predictions, targets)
+    print("sacre blue", sacre_blue_score)
     return blue_score
